% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_building_areas_bins.R
\name{plot_area_bin_comparison}
\alias{plot_area_bin_comparison}
\title{Plot Side-by-Side Area Binned Maps for Structured vs Unstructured Buildings}
\usage{
plot_area_bin_comparison(structured, unstructured)
}
\arguments{
\item{structured}{An sf object with \code{area_bin}.}

\item{unstructured}{An sf object with \code{area_bin}.}
}
\value{
A patchwork ggplot2 object plotting the two maps side-by-side.
}
\description{
Requires the input sf objects to have an \code{area_bin} factor column (use \code{create_bins()} first).
}
