% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geometric_visualize_buildings.R
\name{plot_buildings_overlay}
\alias{plot_buildings_overlay}
\title{Plot overlaid comparison of building geometries}
\usage{
plot_buildings_overlay(buildings)
}
\arguments{
\item{buildings}{A list of two \code{sf} objects: \code{unstructured} and \code{structured}.}
}
\value{
A \code{ggplot} with unstructured buildings in red and structured in blue.
}
\description{
Plot overlaid comparison of building geometries
}
