% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geometric_visualize_buildings.R
\name{plot_buildings_side_by_side}
\alias{plot_buildings_side_by_side}
\title{Plot side-by-side building geometries}
\usage{
plot_buildings_side_by_side(buildings)
}
\arguments{
\item{buildings}{A list of two \code{sf} objects: \code{unstructured} and \code{structured}.}
}
\value{
A \code{ggpatchwork} plot showing both geometries side by side.
}
\description{
Plot side-by-side building geometries
}
