% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_interactive_map.R
\name{plot_interactive_area_map}
\alias{plot_interactive_area_map}
\title{Interactive Area Map for Structured and Unstructured Buildings}
\usage{
plot_interactive_area_map(buildings)
}
\arguments{
\item{buildings}{A list with two \code{sf} objects: \code{structured} and \code{unstructured}.
Each must have a numeric \code{area} column or else it will be computed.}
}
\value{
A \code{tmap} interactive map (in "view" mode).
}
\description{
Creates an interactive leaflet map comparing building areas
for structured and unstructured building datasets.
}
